Tabbed VM Refactor complete
