--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/AboutTab.xaml
@@ -1,53 +1,0 @@
-﻿<UserControl x:Class="ChordFactory.OpenSilver.AboutTab" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:local="clr-namespace:ChordFactory.OpenSilver">
-    <UserControl.Content>
-        <StackPanel Margin="20,0">
-            <StackPanel.Resources>
-                <Style TargetType="StackPanel">
-                    <Setter Property="Margin" Value="0,2" />
-                </Style>
-                <Style TargetType="TextBlock">
-                    <Setter Property="FontSize" Value="20" />
-                </Style>
-            </StackPanel.Resources>
-            <StackPanel.Children>
-
-                <TextBlock x:Name="PageTitle" Margin="0,5,0,8" HorizontalAlignment="Left" FontSize="24" Text="Openfeature ChordFactory" />
-                <StackPanel Margin="0,5,0,8">
-                    <StackPanel.Children>
-                        <TextBlock Text="Copyright" />
-                        <TextBlock Foreground="DarkRed" Text="©2019 Openfeature Limited" />
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Margin="0,5,0,8">
-                    <StackPanel.Children>
-                        <TextBlock Text="Version" />
-                        <TextBlock Foreground="DarkRed" Text="1.0" />
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Margin="0,5,0,8">
-                    <StackPanel.Children>
-                        <TextBlock Text="Contact" />
-                        <TextBlock Foreground="DarkRed" Text="info@openfeature.co.uk" />
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Margin="0,30,20,0" HorizontalAlignment="Left" Orientation="Vertical">
-                    <StackPanel.Children>
-                        <StackPanel>
-                            <StackPanel.Children>
-                                <TextBlock Width="600" HorizontalAlignment="Left" FontSize="22" FontStyle="Italic" Text="This software is a Xamarin version of Openfeature ChordFactory. Gordon Mackie's eternal project to update a simple chord and scale finder utility to every new technology that comes along." />
-                                <TextBlock Margin="0,20,26,20" FontSize="22">
-                                    <TextBlock.Inlines>
-                                        <Run Foreground="Blue" Text="Go to website" TextDecorations="Underline" />
-                                    </TextBlock.Inlines>
-                                    <!--<TextBlock.GestureRecognizers>
-                                        <TapGestureRecognizer Command="{Binding OpenWebCommand}" />
-                                    </TextBlock.GestureRecognizers>-->
-                                </TextBlock>
-                            </StackPanel.Children>
-                        </StackPanel>
-                    </StackPanel.Children>
-                </StackPanel>
-            </StackPanel.Children>
-        </StackPanel>
-    </UserControl.Content>
-</UserControl>
