--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/AboutTab.xaml.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Navigation;
-
-namespace ChordFactory.OpenSilver
-{
-    public partial class AboutTab : UserControl
-    {
-        public AboutTab()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
