+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/App.xaml
@@ -1,11 +1,66 @@
-﻿<Application
-    x:Class="ChordFactory.OpenSilver.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:ChordFactory.OpenSilver">
+﻿<Application x:Class="ChordFactory.OpenSilver.App" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:local="clr-namespace:ChordFactory.OpenSilver">
     <Application.Resources>
         <ResourceDictionary>
-            
+            <Style x:Key="KeyboardGridStyle" TargetType="Grid">
+                <Style.Setters>
+                    <Setter Property="MinHeight" Value="350" />
+                    <Setter Property="MaxHeight" Value="770" />
+                    <Setter Property="HorizontalAlignment" Value="Stretch" />
+                    <Setter Property="VerticalAlignment" Value="Stretch" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="TitleLabel" TargetType="TextBlock">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="32" />
+                    <Setter Property="Foreground" Value="PaleGoldenRod" />
+                    <Setter Property="HorizontalAlignment" Value="Left" />
+                    <Setter Property="Margin" Value="0,0,0,20" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="SelectedItemLabel" TargetType="TextBlock">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="20" />
+                    <Setter Property="Foreground" Value="Orange" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="HeaderLabel" TargetType="TextBlock">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="24" />
+                    <Setter Property="Foreground" Value="Orange" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="ParagraphText" TargetType="TextBlock">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="16" />
+                    <Setter Property="Foreground" Value="LightYellow" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="PianoKeyStyle" TargetType="Border">
+                <Style.Setters>
+                    <Setter Property="BorderBrush" Value="Black" />
+                    <Setter Property="BorderThickness" Value="2" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="ComboTestStyle" TargetType="ComboBox">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="20" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="TabControlStyle" TargetType="TabControl">
+                <Style.Setters>
+                    <Setter Property="FontSize" Value="18" />
+                    <Setter Property="Background" Value="#1E6167" />
+                    <!--<Setter Property="Background" Value="DarkSlateGray" />-->
+                    <Setter Property="HorizontalAlignment" Value="Stretch" />
+                    <Setter Property="VerticalAlignment" Value="Stretch" />
+                </Style.Setters>
+            </Style>
+            <Style x:Key="TabItemStyle" TargetType="TabItem">
+                <Style.Setters>
+                    <Setter Property="Foreground" Value="DarkRed" />
+                    <Setter Property="Background" Value="LightGray" />
+                </Style.Setters>
+            </Style>
         </ResourceDictionary>
     </Application.Resources>
 </Application>
