--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/MainPage - Copy.xaml
@@ -1,118 +1,0 @@
-﻿<Page
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    x:Class="ChordFactory.OpenSilver.MainPage"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:ChordFactory.OpenSilver"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-    <Page.Content>
-        <Grid Background="DarkBlue" Margin="10">
-            <Grid.Resources>
-                <Style x:Key="TextBlockCentred" TargetType="TextBlock">
-                    <Style.Setters>
-                        <Setter Property="HorizontalAlignment" Value="Center" />
-                        <Setter Property="VerticalAlignment" Value="Center" />
-                        <Setter Property="Margin" Value="40,40,0,0" />
-                        <Setter Property="Foreground" Value="PaleGoldenRod" />
-                        <Setter Property="FontSize" Value="40" />
-                        <Setter Property="FontWeight" Value="Bold" />
-                    </Style.Setters>
-                </Style>
-            </Grid.Resources>
-            <Grid.RowDefinitions>
-                <RowDefinition/>
-                <RowDefinition/>
-                <RowDefinition/>
-            </Grid.RowDefinitions>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition/>
-                <ColumnDefinition/>
-                <ColumnDefinition/>
-            </Grid.ColumnDefinitions>
-            <Grid.Children>
-                <StackPanel Grid.Row="0" Grid.Column="0" Background="Red">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Top Left" x:Name="TextBlock1" >
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="0" Grid.Column="1" Background="Orange">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Top Centre" x:Name="TextBlock2">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="0" Grid.Column="2" Background="Yellow">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Top Right" x:Name="TextBlock3">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="1" Grid.Column="0" Background="Green">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Middle Left" x:Name="TextBlock4">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="1" Grid.Column="1" Background="Blue">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Middle Centre" x:Name="TextBlock5">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="1" Grid.Column="2" Background="Indigo">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Middle Right" x:Name="TextBlock6">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="2" Grid.Column="0" Background="Violet">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Bottom Left" x:Name="TextBlock7">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="2" Grid.Column="1" Background="Silver">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Bottom Centre" x:Name="TextBlock8">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-                <StackPanel Grid.Row="2" Grid.Column="2" Background="Black">
-                    <StackPanel.Children>
-                        <TextBlock Style="{StaticResource TextBlockCentred}" Text="Bottom Right" x:Name="TextBlock9">
-                            <TextBlock.RenderTransform>
-                                <RotateTransform Angle="45" />
-                            </TextBlock.RenderTransform>
-                        </TextBlock>
-                    </StackPanel.Children>
-                </StackPanel>
-            </Grid.Children>
-        </Grid>
-    </Page.Content>
-</Page>
