--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/MainPage.xaml
@@ -1,54 +1,0 @@
-﻿<Page x:Class="ChordFactory.OpenSilver.MainPage" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:local="clr-namespace:ChordFactory.OpenSilver" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" Background="DarkSlateGray" mc:Ignorable="d">
-    <Page.Resources>
-        <Style x:Key="TabControlStyle" TargetType="TabControl">
-            <Style.Setters>
-                <Setter Property="FontSize" Value="18" />
-                <Setter Property="Background" Value="Gray" />
-                <Setter Property="HorizontalAlignment" Value="Stretch" />
-                <Setter Property="VerticalAlignment" Value="Stretch" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="TabItemStyle" TargetType="TabItem">
-            <Style.Setters>
-                <Setter Property="Foreground" Value="DarkOrange" />
-                <Setter Property="Background" Value="DarkGray" />
-            </Style.Setters>
-        </Style>
-    </Page.Resources>
-    <Page.Content>
-        <TabControl x:Name="MainPageTabControl" Grid.Row="1" Style="{StaticResource TabControlStyle}">
-            <TabControl.Items>
-                <TabItem HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Header="Chords" IsSelected="True">
-                    <TabItem.Content>
-                        <!--<TextBlock Margin="20" Text="Content of Tab 1" />-->
-                        <local:ChordKeyboardPage HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
-                    </TabItem.Content>
-                </TabItem>
-                <TabItem Header="Scales">
-                    <TabItem.Content>
-                        <!--<TextBlock Margin="20" Text="Content of Tab 2" />-->
-                        <local:ScaleKeyboardPage />
-                    </TabItem.Content>
-                </TabItem>
-                <TabItem Header="Finder" Style="{StaticResource TabItemStyle}">
-                    <TabItem.Content>
-                        <!--<TextBlock Margin="20" Text="Content of Tab 3" />-->
-                        <local:FinderKeyboardPage />
-                    </TabItem.Content>
-                </TabItem>
-                <TabItem Header="Settings">
-                    <TabItem.Content>
-                        <!--<TextBlock Margin="20" Text="Settings Tab" />-->
-                        <local:SettingsPage />
-                    </TabItem.Content>
-                </TabItem>
-                <TabItem Header="About">
-                    <TabItem.Content>
-                        <!--<TextBlock Margin="20" Text="About Tab" />-->
-                        <local:AboutTab />
-                    </TabItem.Content>
-                </TabItem>
-            </TabControl.Items>
-        </TabControl>
-    </Page.Content>
-</Page>
