--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/MainPage.xaml - Copy.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Windows;
-using System.Windows.Controls;
-
-namespace ChordFactory.OpenSilver
-{
-    public partial class MainPage : Page
-    {
-        public MainPage()
-        {
-            this.InitializeComponent();
-
-            // Enter construction logic here...
-        }
-    }
-}
