--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/SettingsPage.xaml
@@ -1,109 +1,0 @@
-﻿<Page
-    x:Class="ChordFactory.OpenSilver.SettingsPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:ChordFactory.OpenSilver" xmlns:controls="clr-namespace:ChordFactory.OpenSilver.controls">
-
-    <Page.Content>
-        <ScrollViewer Margin="20,0,20,0">
-            <ScrollViewer.Content>
-                <Grid>
-                    <Grid.RowDefinitions>
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="Auto" />
-                        <RowDefinition Height="*" />
-                    </Grid.RowDefinitions>
-                    <Grid.Children>
-                        <TextBlock x:Name="PageTitle" Grid.Row="0"
-                   Margin="0,5,0,8" FontSize="24"
-                   HorizontalAlignment="Left" Text="Openfeature ChordFactory" />
-                        <TextBlock Grid.Row="1" Margin="0,2"
-                   FontWeight="Bold" FontSize="16"
-                   HorizontalAlignment="Left" Text="Colours"
-                   Foreground="DarkGray" />
-                        <StackPanel Grid.Row="2" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Chord Black Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="BlackChordColourPicker"
-                                           Colour="{Binding BlackKeySelectedChordColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="3" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Chord White Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="WhiteChordColourPicker"
-                                           Colour="{Binding WhiteKeySelectedChordColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="4" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Scale Black Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="BlackScaleColourPicker"
-                                           Colour="{Binding BlackKeySelectedScaleColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="5" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Scale White Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="WhiteScaleColourPicker"
-                                           Colour="{Binding WhiteKeySelectedScaleColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="6" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Finder Black Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="BlackFinderColourPicker"
-                                           Colour="{Binding BlackKeySelectedFinderColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="7" Margin="0,0,0,2">
-                            <StackPanel.Children>
-                                <TextBlock Text="Finder White Key" VerticalAlignment="Center" />
-                                <controls:ColourPicker x:Name="WhiteFinderColourPicker"
-                                           Colour="{Binding WhiteKeySelectedFinderColour}"
-                                           VerticalAlignment="Center" />
-                            </StackPanel.Children>
-                        </StackPanel>
-                        <StackPanel Grid.Row="8">
-                            <StackPanel.Children>
-                                <TextBlock Margin="0,2" FontWeight="Bold"
-                       FontSize="16" HorizontalAlignment="Left"
-                       Text="Sound" Foreground="DarkGray" />
-                                <StackPanel Margin="0,2,40,2" HorizontalAlignment="Right"
-                             Orientation="Horizontal">
-                                    <StackPanel.Children>
-                                        <TextBlock Text="Play Chords/Scales?" VerticalAlignment="Center" />
-                                        <ToggleButton Margin="10,0,0,0"
-                            IsChecked="{Binding AudioIsEnabled}"
-                            VerticalAlignment="Center" />
-                                    </StackPanel.Children>
-                                </StackPanel>
-                                <StackPanel Margin="0,2,40,2" HorizontalAlignment="Right"
-                             Orientation="Horizontal">
-                                    <StackPanel.Children>
-                                        <TextBlock Text="Arpeggiate?" VerticalAlignment="Center" />
-                                        <ToggleButton Margin="10,0,0,0"
-                            IsEnabled="{Binding AudioIsEnabled}"
-                            IsChecked="{Binding ArpeggiateIsEnabled}"
-                            VerticalAlignment="Center" />
-                                    </StackPanel.Children>
-                                </StackPanel>
-                            </StackPanel.Children>
-                        </StackPanel>
-                    </Grid.Children>
-                </Grid>
-            </ScrollViewer.Content>
-        </ScrollViewer>
-    </Page.Content>
-</Page>
