--- ChordFactory.OpenSilver/ChordFactory.OpenSilver/SettingsPage.xaml.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Controls.Primitives;
-using System.Windows.Data;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Navigation;
-
-namespace ChordFactory.OpenSilver
-{
-    public partial class SettingsPage : Page
-    {
-        public SettingsPage()
-        {
-            this.InitializeComponent();
-        }
-    }
-}
