+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/controls/ColourPicker.xaml
@@ -1,11 +1,5 @@
-﻿<Page
-    x:Class="ChordFactory.OpenSilver.controls.ColourPicker"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:local="clr-namespace:ChordFactory.OpenSilver.controls">
-    <Page.Content>
+﻿<UserControl x:Class="ChordFactory.OpenSilver.controls.ColourPicker" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:local="clr-namespace:ChordFactory.OpenSilver.controls" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006">
+    <UserControl.Content>
         <Grid>
             <Grid.RowDefinitions>
                 <RowDefinition Height="40" />
@@ -15,17 +9,16 @@
                 <ColumnDefinition Width="154" />
             </Grid.ColumnDefinitions>
             <Grid.Children>
-                <Grid Grid.Row="0" Grid.Column="0" Background="Gray"
-                  HorizontalAlignment="Stretch" Margin="2" VerticalAlignment="Stretch">
+                <Grid Grid.Row="0" Grid.Column="0" Margin="2" HorizontalAlignment="Stretch"
+                      VerticalAlignment="Stretch" Background="Gray">
                     <Grid.Children>
-                        <Border x:Name="ColourPickerBoxView" BorderBrush="{Binding Colour}" />
+                        <Border x:Name="ColourPickerBoxView" Background="{Binding Colour}" />
                     </Grid.Children>
                 </Grid>
-                <Image x:Name="ColourPickerImage" Grid.Row="0" Grid.Column="0"
-                   HorizontalAlignment="Stretch" Source="ColourPicker.Assets.ColourPicker.png" VerticalAlignment="Stretch" />
-                <ComboBox x:Name="SelectionPicker" Grid.Row="0" Grid.Column="1"
-                    SelectionChanged="PickerSelectedIndexChanged" />
+                <!--<Image x:Name="ColourPickerImage" Grid.Row="0" Grid.Column="0" HorizontalAlignment="Stretch"
+                       VerticalAlignment="Stretch" Source="images.ColourPicker.png" />-->
+                <ComboBox x:Name="SelectionPicker" Grid.Row="0" Grid.Column="1" SelectionChanged="PickerSelectedIndexChanged" />
             </Grid.Children>
         </Grid>
-    </Page.Content>
-</Page>
+    </UserControl.Content>
+</UserControl>
