+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/viewModels/ChordKeyboardViewModel.cs
@@ -1,9 +1,9 @@
 ﻿namespace ChordFactory.OpenSilver.viewModels
 {
     using System;
+    using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Diagnostics;
-
     using models;
 
     public class ChordKeyboardViewModel : BaseViewModel
@@ -13,28 +13,27 @@
             this.Title = "Keyboard";
             this.Items = new ObservableCollection<Item>();
             this.LoadItemsCommand = new DelegateCommand(this.ExecuteLoadItemsCommand(), this.ExecuteLoadItemsCanExecute);
-
-            //MessagingCenter.Subscribe<NewItemPage, Item>(
-            //    this, "AddItem", async (obj, item) =>
-            //    {
-            //        var newItem = item;
-            //        this.Items.Add(newItem);
-            //        await this.DataStore.AddItemAsync(newItem);
-            //    });
         }
+        
+        public List<Chord> Chords => this.MusicData.Chords;
+
+        public string[] Inversions => this.MusicData.Inversions;
+
+        public ObservableCollection<Item> Items { get; set; }
+
+        public DelegateCommand LoadItemsCommand { get; set; }
 
         private bool ExecuteLoadItemsCanExecute(object paramList)
         {
             return true;
         }
 
-        public ObservableCollection<Item> Items { get; set; }
-        public DelegateCommand LoadItemsCommand { get; set; }
-
         private Action<object> ExecuteLoadItemsCommand()
         {
             if (this.IsBusy)
+            {
                 return null;
+            }
 
             this.IsBusy = true;
 
@@ -54,7 +53,6 @@
             finally
             {
                 this.IsBusy = false;
-
             }
 
             return null;
