+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/AboutControl.xaml
@@ -1,0 +1,48 @@
+﻿<UserControl x:Class="ChordFactory.OpenSilver.views.AboutControl" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:local="clr-namespace:ChordFactory.OpenSilver">
+    <UserControl.Content>
+        <StackPanel Margin="30" HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
+            <StackPanel.Resources>
+                <Style TargetType="StackPanel">
+                    <Setter Property="Margin" Value="0,2" />
+                </Style>
+                <Style TargetType="TextBlock">
+                    <Setter Property="FontSize" Value="20" />
+                </Style>
+            </StackPanel.Resources>
+            <StackPanel.Children>
+
+                <TextBlock x:Name="PageTitle" Style="{StaticResource TitleLabel}" Text="Openfeature ChordFactory" />
+                <StackPanel Margin="0,5,0,8">
+                    <StackPanel.Children>
+                        <TextBlock Style="{StaticResource HeaderLabel}" Text="Copyright" />
+                        <TextBlock Style="{StaticResource ParagraphText}" Text="©2019 Openfeature Limited" />
+                    </StackPanel.Children>
+                </StackPanel>
+                <StackPanel Margin="0,5,0,8">
+                    <StackPanel.Children>
+                        <TextBlock Style="{StaticResource HeaderLabel}" Text="Version" />
+                        <TextBlock Style="{StaticResource ParagraphText}" Text="1.0" />
+                    </StackPanel.Children>
+                </StackPanel>
+                <StackPanel Margin="0,5,0,8">
+                    <StackPanel.Children>
+                        <TextBlock Style="{StaticResource HeaderLabel}" Text="Contact" />
+                        <TextBlock Style="{StaticResource ParagraphText}" Text="info@openfeature.co.uk" />
+                    </StackPanel.Children>
+                </StackPanel>
+                <StackPanel Margin="0,30,20,0" HorizontalAlignment="Left" Orientation="Vertical">
+                    <StackPanel.Children>
+                        <StackPanel>
+                            <StackPanel.Children>
+                                <TextBlock Width="600" HorizontalAlignment="Left" FontSize="22" FontStyle="Italic"
+                                           Style="{StaticResource ParagraphText}"
+                                           Text="This software is a Xamarin version of Openfeature ChordFactory. Gordon Mackie's eternal project to update a simple chord and scale finder utility to every new technology that comes along." TextWrapping="Wrap" />
+                                <HyperlinkButton Margin="0,20" HorizontalAlignment="Left" FontSize="22" Content="More information..." TargetName="_blank" NavigateUri="http://www.openfeature.com" />
+                            </StackPanel.Children>
+                        </StackPanel>
+                    </StackPanel.Children>
+                </StackPanel>
+            </StackPanel.Children>
+        </StackPanel>
+    </UserControl.Content>
+</UserControl>
