+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/AboutControl.xaml.cs
@@ -1,0 +1,12 @@
+﻿namespace ChordFactory.OpenSilver.views
+{
+    using System.Windows.Controls;
+
+    public partial class AboutControl : UserControl
+    {
+        public AboutControl()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
