+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/ChordKeyboardControl.xaml
@@ -1,45 +1,5 @@
-﻿<Page x:Class="ChordFactory.OpenSilver.FinderKeyboardPage" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Background="Transparent" mc:Ignorable="d">
-    <Page.Resources>
-        <Style x:Key="KeyboardGridStyle" TargetType="Grid">
-            <Style.Setters>
-                <Setter Property="MinHeight" Value="350" />
-                <Setter Property="MaxHeight" Value="770" />
-                <Setter Property="HorizontalAlignment" Value="Stretch" />
-                <Setter Property="VerticalAlignment" Value="Stretch" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="TitleLabel" TargetType="TextBlock">
-            <Style.Setters>
-                <Setter Property="FontSize" Value="32" />
-                <Setter Property="Foreground" Value="PaleGoldenRod" />
-                <Setter Property="Margin" Value="0,0,0,20" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="SelectedItemLabel" TargetType="TextBlock">
-            <Style.Setters>
-                <Setter Property="FontSize" Value="24" />
-                <Setter Property="Foreground" Value="Orange" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="HeaderLabel" TargetType="TextBlock">
-            <Style.Setters>
-                <Setter Property="FontSize" Value="16" />
-                <Setter Property="Foreground" Value="Orange" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="PianoKeyStyle" TargetType="Border">
-            <Style.Setters>
-                <Setter Property="BorderBrush" Value="Black" />
-                <Setter Property="BorderThickness" Value="2" />
-            </Style.Setters>
-        </Style>
-        <Style x:Key="ComboTestStyle" TargetType="ComboBox">
-            <Style.Setters>
-                <Setter Property="FontSize" Value="20" />
-            </Style.Setters>
-        </Style>
-    </Page.Resources>
-    <Page.Content>
+﻿<UserControl x:Class="ChordFactory.OpenSilver.views.ChordKeyboardControl" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" Background="Transparent" mc:Ignorable="d">
+    <UserControl.Content>
         <Grid x:Name="LayoutRoot" Margin="30" HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto" />
@@ -49,10 +9,32 @@
             </Grid.RowDefinitions>
             <Grid.Children>
                 <TextBlock Grid.Row="0" Style="{StaticResource TitleLabel}" Text="Openfeature ChordFactory" />
-                <TextBlock Grid.Row="1" Style="{StaticResource HeaderLabel}" Text="Chord Finder" />
+                <Grid Grid.Row="1">
+                    <Grid.RowDefinitions>
+                        <RowDefinition />
+                        <RowDefinition />
+                    </Grid.RowDefinitions>
+                    <Grid.ColumnDefinitions>
+                        <ColumnDefinition />
+                        <ColumnDefinition />
+                    </Grid.ColumnDefinitions>
+                    <Grid.Children>
+                        <TextBlock Grid.Row="0" Grid.Column="0" Style="{StaticResource HeaderLabel}" Text="Chords" />
+                        <TextBlock Grid.Row="0" Grid.Column="1" Margin="0,0,30,0" HorizontalAlignment="Right"
+                                   Style="{StaticResource HeaderLabel}"
+                                   Text="Inversion" />
+                        <ComboBox x:Name="ChordsComboBox" Grid.Row="1" Grid.Column="0" Margin="0,10"
+                                  HorizontalAlignment="Left" DisplayMemberPath="Description"
+                                  ItemsSource="{Binding Chords}"
+                                  SelectedIndex="0" />
+                        <ComboBox x:Name="InversionCombo" Grid.Row="1" Grid.Column="1" Margin="0,10,30,10"
+                                  HorizontalAlignment="Right"
+                                  ItemsSource="{Binding Inversions}"
+                                  SelectedIndex="0" />
+                    </Grid.Children>
+                </Grid>
                 <Grid x:Name="ChordKeyboardGrid" Grid.Row="2" Margin="0,10,30,10" HorizontalAlignment="Stretch"
-                      VerticalAlignment="Top" Background="Ivory"
-                      Style="{StaticResource KeyboardGridStyle}">
+                      VerticalAlignment="Top" Background="Ivory">
                     <Grid.ColumnDefinitions>
                         <ColumnDefinition Width="*" />
                         <ColumnDefinition Width="*" />
@@ -214,10 +196,12 @@
                     <TextBlock.Inlines>
                         <Run Text="Current Chord" />
                         <LineBreak />
-                        <Run x:Name="SelectedChordLabel" FontStyle="Italic" Foreground="Yellow" Text="{Binding FoundChordFullName}" />
+                        <Run x:Name="SelectedChordLabel" FontStyle="Italic" Foreground="Yellow" />
+                        <LineBreak />
+                        <Run x:Name="SelectedChordInversionNotesLabel" FontStyle="Italic" Foreground="LightBlue" />
                     </TextBlock.Inlines>
                 </TextBlock>
             </Grid.Children>
         </Grid>
-    </Page.Content>
-</Page>
+    </UserControl.Content>
+</UserControl>
