+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/FinderKeyboardControl.xaml.cs
@@ -1,4 +1,4 @@
-﻿namespace ChordFactory.OpenSilver
+﻿namespace ChordFactory.OpenSilver.views
 {
     using System;
     using System.Collections.Generic;
@@ -6,14 +6,12 @@
     using System.Linq;
     using System.Windows;
     using System.Windows.Controls;
-    using System.Windows.Documents;
     using System.Windows.Input;
     using System.Windows.Media;
-
     using models;
+    using viewModels;
 
-    
-    public partial class ChordKeyboardPage : Page, INotifyPropertyChanged
+    public partial class FinderKeyboardControl : UserControl, INotifyPropertyChanged
     {
         //private const string WaveString = @"ChordFactory.OpenSilver;resources/media/PIANO_MED_{0}.mp3";
         private const string WaveString = @"ms-appx:///media/PIANO_MED_{0}.mp3";
@@ -34,51 +32,25 @@
         };
 
         private Grid chordKeyboardGrid;
-
         private int chordRootNote;
-        private ComboBox chordsCombo;
 
-        private int finderRootNoteOffset;
-        private ComboBox inversionCombo;
-        private Run selectedChordInversionNotesLabel;
-        private Run selectedChordLabel;
-
-        public ChordKeyboardPage()
+        public FinderKeyboardControl()
         {
             this.InitializeComponent();
             this.Loaded += this.MainPage_Loaded;
-            this.PropertyChanged += this.MainPage_PropertyChanged;
-
         }
 
-        private void MainPage_PropertyChanged(object sender, PropertyChangedEventArgs e)
-        {
-        }
-
-        public MusicData MusicData { get; set; }
+        public FinderViewModel FinderViewModel => this.DataContext as FinderViewModel;
 
         public InversionEnum SelectedInversion { get; set; }
 
+        public Chord IdentifiedChord { get; set; }
+
         private void MainPage_Loaded(object sender, RoutedEventArgs e)
         {
-            this.MusicData = new MusicData();
-            this.DataContext = this.MusicData;
-            this.chordsCombo = this.FindName("ChordsComboBox") as ComboBox;
-            this.selectedChordLabel = this.FindName("SelectedChordLabel") as Run;
-            this.selectedChordInversionNotesLabel = this.FindName("SelectedChordInversionNotesLabel") as Run;
-
-
-            this.inversionCombo = this.FindName("InversionCombo") as ComboBox;
-
-            if (this.inversionCombo != null)
-            {
-                this.inversionCombo.SelectionChanged += this.InversionCombo_SelectionChanged;
-            }
-
-            if (this.chordsCombo != null)
-            {
-                this.chordsCombo.SelectionChanged += this.ChordsCombo_SelectionChanged;
-            }
+            this.DataContext = ((App)Application.Current)?.FinderViewModel;
+            this.FinderViewModel.PropertyChanged += this.FinderViewModelPropertyChanged;
+            this.FinderViewModel.Settings.PropertyChanged += this.SettingsPropertyChanged;
 
             this.chordKeyboardGrid = (Grid)this.FindName("ChordKeyboardGrid");
 
@@ -89,18 +61,23 @@
 
             this.PopulateKeys();
 
-            if (this.chordsCombo.Items.Count > 0)
-            {
-                this.chordsCombo.SelectedIndex = 0;
-            }
-
-            if (this.inversionCombo.Items.Count > 0)
-            {
-                this.inversionCombo.SelectedIndex = 0;
-            }
-
             this.SelectedInversion = 0;
             this.AdjustKeyboardAspectRatios();
+        }
+
+        private void SettingsPropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+
+
+        }
+
+        private void FinderViewModelPropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            if (e.PropertyName == "FinderRootNoteOffset")
+            {
+                this.ShowChord();
+                // this.PlayChord();
+            }
         }
 
         private void MainPage_SizeChanged(object sender, SizeChangedEventArgs e)
@@ -111,16 +88,7 @@
         private void AdjustKeyboardAspectRatios()
         {
             var newKeyboardHeight = this.chordKeyboardGrid.ActualWidth * 0.45;
-            this.chordKeyboardGrid.Height = Math.Min(400, Math.Max(newKeyboardHeight, 300));
-        }
-
-        private void InversionCombo_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            this.SelectedInversion = (InversionEnum)((ComboBox)sender).SelectedIndex;
-            if (this.chordsCombo.SelectedItem != null)
-            {
-                this.ShowChord(this.chordsCombo.SelectedItem as Chord);
-            }
+            this.chordKeyboardGrid.Height = newKeyboardHeight;
         }
 
         private void PopulateKeys()
@@ -148,12 +116,6 @@
             var pauseLength = new TimeSpan(0, 0, 0, 1, 500);
 
             var actualChordNotes = chord.Notes.Select(c => (c + this.chordRootNote) % 24).ToList();
-            for (var inversionNote = 0;
-                inversionNote < Math.Min(this.inversionCombo.SelectedIndex, chord.Notes.Count);
-                inversionNote++)
-            {
-                actualChordNotes[inversionNote] += 12;
-            }
 
             actualChordNotes.Sort();
 
@@ -187,58 +149,34 @@
         private void ChordKey_Tapped(object sender, TappedRoutedEventArgs e)
         {
             this.chordRootNote = this.chordKeys.IndexOf(sender as Border) % 12;
-            this.ShowChord(this.chordsCombo.SelectedItem as Chord);
+            this.FinderViewModel.FinderKeyboardTappedCommand.Execute($"Key{this.chordKeys.IndexOf(sender as Border)}");
+            this.ShowChord();
         }
 
-        private void ChordsCombo_SelectionChanged(object sender, SelectionChangedEventArgs e)
-        {
-            if ((sender as ComboBox)?.SelectedItem is Chord chord)
-            {
-                this.ShowChord(chord);
-            }
-        }
-
-        private void ShowChord(Chord chord)
+        private void ShowChord()
         {
             this.ClearKeySelection();
 
-            var adjustedNotes = new int[chord.Notes.Count];
-            chord.Notes.CopyTo(adjustedNotes);
-
-            for (var inversionNote = 0;
-                inversionNote < Math.Min((int)this.SelectedInversion, chord.Notes.Count);
-                inversionNote++)
+            if (this.FinderViewModel.FinderChord.Notes.Count > 0)
             {
-                adjustedNotes[inversionNote] += 12;
+                var actualChordNotes = this.FinderViewModel.FinderChord.Notes.ToArray();
+
+                // Cope with overflow
+                for (var note = 0; note < actualChordNotes.Length; note++)
+                {
+                    actualChordNotes[note] = actualChordNotes[note] % this.chordKeys.Count;
+                }
+
+                foreach (var note in actualChordNotes)
+                {
+                    var colourTag = (string)this.chordKeys[note].Tag;
+                    this.chordKeys[note].Background =
+                        new SolidColorBrush(colourTag == "Ivory" ? this.chordWhiteKeySelected : this.chordBlackKeySelected);
+                    this.chordKeys[note].BorderBrush = new SolidColorBrush(this.chordKeyBorderSelected);
+                }
             }
-
-            for (var index = 0; index < adjustedNotes.Length; index++)
-            {
-                adjustedNotes[index] = adjustedNotes[index] % 24;
-            }
-
-            foreach (var note in adjustedNotes)
-            {
-                var adjustedNoteIndex = (note + this.chordRootNote) % 24;
-                var colourTag = (string)this.chordKeys[adjustedNoteIndex].Tag;
-                this.chordKeys[adjustedNoteIndex].Background =
-                    new SolidColorBrush(colourTag == "Ivory" ? this.chordWhiteKeySelected : this.chordBlackKeySelected);
-                this.chordKeys[adjustedNoteIndex].BorderBrush = new SolidColorBrush(this.chordKeyBorderSelected);
-            }
-
-            this.selectedChordLabel.Text = $"{this.noteNames[this.chordRootNote]} {chord.Description}";
-
-            var inversionText = string.Empty;
-            if (this.SelectedInversion != 0)
-            {
-                inversionText = $" - {this.SelectedInversion.ToString().ToLower()} inversion";
-            }
-
-            this.selectedChordInversionNotesLabel.Text = $"[{this.GetChordNoteNamesText(chord.Notes)}{inversionText}]";
-
-            // this.PlayChord(chord);
         }
-        
+
         private void ClearKeySelection()
         {
             this.chordKeys.ForEach(
@@ -267,13 +205,6 @@
         {
             var adjustedNotes = new List<int>(noteSequence);
             var adjustedNoteNames = new string[noteSequence.Count];
-
-            for (var inversionNote = 0;
-                inversionNote < Math.Min(this.inversionCombo.SelectedIndex, noteSequence.Count);
-                inversionNote++)
-            {
-                adjustedNotes[inversionNote] += 12;
-            }
 
             adjustedNotes.Sort();
