+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/MainPage.xaml
@@ -1,0 +1,38 @@
+﻿<Page x:Class="ChordFactory.OpenSilver.views.MainPage" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:local="clr-namespace:ChordFactory.OpenSilver.views" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" Background="DarkSlateGray" mc:Ignorable="d">
+    <Page.Content>
+        <TabControl x:Name="MainPageTabControl" Grid.Row="1" Style="{StaticResource TabControlStyle}">
+            <TabControl.Items>
+                <TabItem HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Header="Chords" IsSelected="True" Style="{StaticResource TabItemStyle}">
+                    <TabItem.Content>
+                        <!--<TextBlock Margin="20" Text="Content of Tab 1" />-->
+                        <local:ChordKeyboardControl HorizontalAlignment="Stretch" VerticalAlignment="Stretch" />
+                    </TabItem.Content>
+                </TabItem>
+                <TabItem Header="Scales" Style="{StaticResource TabItemStyle}">
+                    <TabItem.Content>
+                        <!--<TextBlock Margin="20" Text="Content of Tab 2" />-->
+                        <local:ScaleKeyboardControl />
+                    </TabItem.Content>
+                </TabItem>
+                <TabItem Header="Finder" Style="{StaticResource TabItemStyle}">
+                    <TabItem.Content>
+                        <!--<TextBlock Margin="20" Text="Content of Tab 3" />-->
+                        <local:FinderKeyboardControl />
+                    </TabItem.Content>
+                </TabItem>
+                <TabItem Header="Settings" Style="{StaticResource TabItemStyle}">
+                    <TabItem.Content>
+                        <!--<TextBlock Margin="20" Text="Settings Tab" />-->
+                        <local:SettingsControl />
+                    </TabItem.Content>
+                </TabItem>
+                <TabItem Header="About" Style="{StaticResource TabItemStyle}">
+                    <TabItem.Content>
+                        <!--<TextBlock Margin="20" Text="About Tab" />-->
+                        <local:AboutControl />
+                    </TabItem.Content>
+                </TabItem>
+            </TabControl.Items>
+        </TabControl>
+    </Page.Content>
+</Page>
