+++ ChordFactory.OpenSilver/ChordFactory.OpenSilver/views/SettingsControl.xaml.cs
@@ -1,0 +1,12 @@
+﻿namespace ChordFactory.OpenSilver.views
+{
+    using System.Windows.Controls;
+
+    public partial class SettingsControl : UserControl
+    {
+        public SettingsControl()
+        {
+            this.InitializeComponent();
+        }
+    }
+}
